refactor: create separated compopent for auth-form and use it on the home page as well as on the signin
