+++ src/app/components/auth-form/app.auth-form.scss
@@ -1,0 +1,142 @@
+@mixin border-with-shadow {
+  border-radius: 10px;
+  box-shadow: 2px 4px 7px -2px rgba(34, 60, 80, 0.2);
+}
+
+@mixin error {
+  opacity: 0;
+  font-size: 14px;
+  color: #FE8366;
+  transition: .3s;
+}
+
+@mixin disabledAhref {
+  pointer-events: none;
+  cursor: default;
+  background-color: #b9b9b9;
+  border-color: #b9b9b9;
+  
+  &:hover {
+    color: white;
+  } 
+}
+
+.registration-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  min-height: calc(100vh - 300px);
+  margin-top: 50px;
+
+  img {
+    width: 25vw;
+
+    @media (max-width: 780px) {
+      display: none;
+    }
+  }
+
+  .registration-form {
+    width: 660px;
+    background-color: #E5E5E5;
+    padding: 30px 60px;
+    margin: 0 0 50px 100px;
+    @include border-with-shadow;
+
+    @media (max-width: 780px) {
+      margin: 0;
+    }
+
+    .errorMsg {
+      @include error;
+      display: block;
+      text-align: center;
+      margin-bottom: 10px;
+    }
+
+    h3 {
+      margin-bottom: 10px;
+    }
+
+    .form-fields {
+
+      input {
+        width: 96%;
+        border: none;
+        font-size: 22px;
+        line-height: 28px;
+        padding: 15px 20px;
+        border-radius: 10px;
+
+        @media (max-width: 780px) {
+          font-size: 16px;
+          padding: 5px 10px;
+          margin-bottom: 15px;
+        }
+      }
+
+      .field {
+        margin: 30px 0;
+        position: relative;
+
+        span {
+          position: absolute;
+          bottom: -20px;
+          left: 5px;
+          @include error;
+        }
+      }
+    }
+
+    .registration-block {
+      display: none;
+      flex-direction: column;
+      align-items: center;
+
+      .registration-btn {
+        margin: 20px 0 20px;
+        padding: 15px 40px;
+      }
+
+      .disabled {
+        @include disabledAhref;
+      }
+    }
+
+    .authorization-block {
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+
+      .authorization-btn {
+        display: block;
+        margin: 20px 0 20px;
+        width: 170px;
+        padding: 15px 0;
+
+        @media (max-width: 780px) {
+          padding: 10px 30px;
+        }
+
+        &:hover {
+          svg path {
+            fill: white;
+          }
+        }
+      }
+
+      .disabled {
+        @include disabledAhref;
+        color: white;
+      }
+    }
+
+    .link-btn {
+      border: none;
+      outline: none;
+      background-color: transparent;
+      text-decoration: underline;
+      cursor: pointer;
+    }
+  }
+}
