+++ src/app/pages/home-page/home-page.component.html
@@ -94,30 +94,7 @@
         </div>
     </section>
     <section class="registration">
-        <div class="container registration-container">
-            <img src="../../../assets/free-trial-img.png" alt="">
-            <form class="registration-form">
-                <div class="block-headers">
-                    <h2>Готов начать обучение?</h2>
-                    <h3>Регистрируйся! Это бесплатно!</h3>
-                </div>
-                <div class="form-fields">
-                    <input type="text" id="name" name="username" placeholder="Имя">
-                    <input type="email" name="email" required placeholder="E-mail*">
-                    <input type="password" id="pwd" name="password" required placeholder="Пароль*">
-                </div>
-                <p>* - поля, обязательные для заполнения</p>
-                <div class="form-btns">
-                    <button class="registration-btn fill-btn" type="submit">Зарегистрироваться</button>
-                    <a href="#auth" class="empty-btn authorization-btn">
-                        <svg width="54" height="24" viewBox="0 0 54 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-                            <path d="M52.6624 13.0607C53.2482 12.4749 53.2482 11.5251 52.6624 10.9393L43.1165 1.3934C42.5307 0.807611 41.581 0.807611 40.9952 1.3934C40.4094 1.97919 40.4094 2.92893 40.9952 3.51472L49.4804 12L40.9952 20.4853C40.4094 21.0711 40.4094 22.0208 40.9952 22.6066C41.581 23.1924 42.5307 23.1924 43.1165 22.6066L52.6624 13.0607ZM0 13.5H51.6018V10.5H0V13.5Z" fill="#FE8366"/>
-                        </svg>
-                        Войти
-                    </a>
-                </div>
-            </form>
-        </div>
+        <app-auth-form></app-auth-form>
     </section>
 </main>
 <app-footer></app-footer>
