+++ src/app/pages/home-page/home-page.component.scss
@@ -600,113 +600,6 @@ h3 {
   }
 
   .registration-container {
-    @include flex-between;
-    justify-content: center;
-    z-index: 1;
     margin-top: -100px;
-    margin-bottom: 50px;
-
-    @media (max-width: 560px) {
-      margin-top: 0;
-    }
-
-    img {
-      width: 30vw;
-
-      @media (max-width: 780px) {
-        display: none;
-      }
-    }
-
-    .registration-form {
-      width: 660px;
-      background-color: #E5E5E5;
-      padding: 30px 60px;
-      margin-left: 100px;
-      @include border-with-shadow;
-
-      @media (max-width: 1048px) {
-        width: 300px;
-        margin-left: 50px;
-        padding: 20px 40px;
-
-        h3 {
-          margin-bottom: 20px;
-        }
-      }
-
-      @media (max-width: 780px) {
-        margin-left: 0;
-      }
-
-      input {
-        width: 96%;
-        border: none;
-        font-size: 22px;
-        line-height: 28px;
-        padding: 15px 20px;
-        border-radius: 10px;
-        margin-bottom: 30px;
-
-        @media (max-width: 1048px) {
-          width: 89%;
-        }
-
-        @media (max-width: 780px) {
-          font-size: 16px;
-          padding: 5px 10px;
-          margin-bottom: 15px;
-        }
-      }
-
-      .form-btns {
-        @include flex-between;
-        margin-top: 30px;
-
-        @media (max-width: 1048px) {
-          flex-wrap: wrap;
-        }
-
-        .authorization-btn {
-          width: 220px;
-          padding: 15px 0;
-
-          @media (max-width: 1440px) {
-            width: 180px;
-            padding: 10px 0;
-          }
-
-          @media (max-width: 1048px) {
-            width: 100%;
-          }
-        
-          &:hover {
-        
-            svg path {
-              fill: white;
-            }
-          }
-        }
-
-        .registration-btn {
-          padding: 15px 40px;
-
-          @media (max-width: 1440px) {
-            padding: 14px 40px;
-            margin-right: 30px;
-          }
-
-          @media (max-width: 1048px) {
-            width: 100%;
-            margin-right: 0;
-            margin-bottom: 10px;
-          }
-
-          @media (max-width: 780px) {
-            padding: 10px 0;
-          }
-        }
-      }
-    }
   }
 }
