+++ src/index.ts
@@ -3,17 +3,3 @@ import { appModule } from './app/app.module';
 import './main.scss';
 
 bootstrap(appModule);
-
-/* Passwords
-
-atlas
-vy7Z1OQ88wEVvuOb
-
-heroku
-m0Ol6P9JstR%
-
-mongodb+srv://tanyasamal:vy7Z1OQ88wEVvuOb@cluster0.a1ncc.mongodb.net/learnwords?retryWrites=true&w=majority
-
-JWT_SECRET_KEY lHl1NedCr3a!
-
-JWT_REFRESH_SECRET_KEY ro85SS4UkFUu */
